move cli code to lib.rs Move code to lib.rs, to improve testability
